requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2
django-cors-headers==4.6.0
django-crontab==0.7.1
django-debug-toolbar==5.2.0
django-filter==25.1
django-forms-bootstrap==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
nextsms==0.4
openpyxl==3.1.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.24.0
selcom-apigw-client==1.0.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uuid==1.30
